Move functions under api path
